Add password encryption
Remove the Encoder class as it is not needed, and use BCrypt to encrypt passwords in the UsersController. Set up the security configuration to disable csrf for now to allow post requests with basic auth through.
