+++ application.properties
@@ -2,3 +2,6 @@ server.port = 8080
 
 spring.data.mongodb.host=localhost
 spring.data.mongodb.database=pacman
+logging.level.org.springframework.security=DEBUG
+spring.security.user.name=user
+spring.security.user.password=pacman
