+++ src/main/java/pacmanapi/config/Security.java
@@ -1,0 +1,25 @@
+package pacmanapi.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+import static org.springframework.security.config.Customizer.withDefaults;
+
+@Configuration
+@EnableWebSecurity
+public class Security {
+  @Bean
+  public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+    http
+            .authorizeHttpRequests((authorize) -> authorize
+                    .anyRequest().authenticated()
+            )
+            .httpBasic(withDefaults())
+            .formLogin(withDefaults());
+    http.csrf().disable();
+    return http.build();
+  }
+}
